backend/requirements.txt
fastapi
uvicorn
weaviate-client
sentence-transformers
beautifulsoup4
requests 